Displaying space instead of line and paragraph separators
Chrome renders line separator (0x2028) and paragraph separator (0x2029)
as either a missing glyph or l-sep/p-sep symbol on various platforms.

On Android this happens due to the lack of the glyphs in Android system
fonts, so it is displayed as a missing glyph. Windows system fonts have
l-sep and p-sep symbols for the 0x2028 and 0x2029 codepoints
respectively. On Mac OS these glyphs are being synthesized, see [0].

According to unicode specification line separator and paragraph
separator should be displayed as space, compare [1].

[0] https://crrev.com/c/4180727
[1] https://www.unicode.org/faq/unsup_char.html#2.

Bug: 550275
Change-Id: Ifd472f591daddfa9e6543c37d0a8220c67108592