COOP: restrict-properties 9/*: Block named targeting across browsing context groups.
This patch disables named targeting between frames not belonging to the
same browsing context group in the same CoopRelatedGroup. This is done
in LocalFrame::CanNavigate by checking the two Frame's Page's browsing
context group tokens and CoopRelatedGroup tokens. This path is used for
all target resolution based on name, be it window.open(), anchor target,
form targets, etc.

The state of named targeting and name availability after this patch:
- Frame targeting is blocked in the renderer.
- Frame navigations, more specifically, were already checked in the
browser process in RenderFrameProxyHost::OpenURL.
- Name updates are still carried out to all existing proxies. Proxies
between browsing context groups are limited to the direct group boundary
(the opener or the openee). Therefore we leak name updates of these
boundaries to a compromised renderer.

Further leak reduction is planned, tracked here:
https://bugs.chromium.org/p/chromium/issues/detail?id=1370357

This patch also corrects tests that were outdated after we decided to
clear names when swapping browsing context groups in the same
CoopRelatedGroup.

Bug: 1221127, 1370350
Change-Id: Ieb7b55d02dc266613b45b54299e8e9798462f4eb